fix(test): moved test to bash only, as jsh does not support that kind of parsing

+++ tests/test_redirection.c
@@ -29,6 +29,9 @@ test_info *test_redirection() {
 }
 
 void test_redirection_stdin(test_info *info) {
+    if (!allow_slow) {
+        return;
+    }
 
     print_test_name("Testing input redirection");
 
@@ -43,7 +46,10 @@ void test_redirection_stdin(test_info *info) {
     command *command = parse_command("cat");
     command->call->stdin = fd;
     command->call->stdout = output_fd;
+    command->call->background = 1;
     command_result *result = execute_command(command);
+
+    sleep(1);
 
     destroy_command_result(result);
 
@@ -58,15 +64,18 @@ void test_redirection_stdin(test_info *info) {
 }
 
 void test_redirection_stdout(test_info *info) {
-
+    if (!allow_slow) {
+        return;
+    }
     print_test_name("Testing output redirection");
 
     int fd = open_test_file_to_write("test_redirection_output.log");
 
     command *command = parse_command("echo HelloWorld");
     command->call->stdout = fd;
+    command->call->background = 1;
     command_result *result = execute_command(command);
-
+    sleep(1);
     destroy_command_result(result);
 
     int read_fd = open_test_file_to_read("test_redirection_output.log");
@@ -81,15 +90,18 @@ void test_redirection_stdout(test_info *info) {
 }
 
 void test_redirection_stderr(test_info *info) {
-
+    if (!allow_slow) {
+        return;
+    }
     print_test_name("Testing output redirection");
 
     int fd = open_test_file_to_write("test_redirection_error.log");
 
     command *command = parse_command("mv .");
     command->call->stderr = fd;
+    command->call->background = 1;
     command_result *result = execute_command(command);
-
+    sleep(1);
     destroy_command_result(result);
 
     int read_fd = open_test_file_to_read("test_redirection_output.log");
